Changing issue_tracker model to user Project model
THIS CHANGE REQUIRES DB RESET.

We are finally changing from a static list of projects
to using the requirements tracker's project model. As a
part of this change, the demo data command was updated
to add project population as part of the steps, so new issues
can optionally be setup with a project.

Given our tests are not yet transitioned over, there might be
latent bugs. I did a solid manual pass and found nothing.