Demo data more constrained.
The demo data generation has been constrained further. Assigneed
and verifier will now only populate if they are a part of the
given project. As well, projects now have a random 1-5 developers
added to them.

This was done because if issue_tracker links to a specific project,
then the user might not have access to that project.